Add alloc use
