+++ dds/src/rtps/mod.rs
@@ -12,3 +12,4 @@ pub mod stateless_writer;
 pub mod types;
 pub mod udp_transport;
 pub mod writer_proxy;
+pub mod message_creator;
