+++ dds/src/transport/writer.rs
@@ -2,6 +2,7 @@ use super::{
     history_cache::HistoryCache,
     types::{DurabilityKind, EntityId, Guid, Locator, ReliabilityKind},
 };
+use alloc::vec::Vec;
 
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub struct ReaderProxy {
